fix: konnector status in intent
